Service templates rest tests
